Object 변환기능을 ConvertionHelper로 이동
